move shared files to a new shared folder
